Handle failures in websocket on wincairo
Add more error checks to places which can fail on windows when a
websocket is used. Notably when the connection is lost or when the the
data to send is an even number in the buffer, which requires looping to
send it all. (The socket only signals once that it is ready for more
data to be sent).

Also make sure these errors are reported, instead of silently failing.